+++ src/PokeAByte.Domain/Exceptions/VisibleException.cs
@@ -6,3 +6,13 @@ public class VisibleException : Exception
     {
     }
 }
+
+/// <summary>
+/// Thrown when the mapper XML failed to load for some reason.
+/// </summary>
+public class MapperException : Exception
+{
+    public MapperException(string message, Exception? innerException = null) : base(message, innerException)
+    {
+    }
+}